Attempted fix for coverage>=4.4 filepath changes
This is a partial fix for the issue discussed here:
https://github.com/codeclimate/python-test-reporter/issues/40

The Formatter class now checks the "sources" element against the
current working directory. If these are different the relative path
is prepended to the file_node filename.

This breaks backwards compatibility with coverage<4.4.

It will also not work if there are multiple source directories, as these are
not presently relatable to specific files.
See https://bitbucket.org/ned/coveragepy/issues/578/incomplete-file-path-in-xml-report